Refactor release candidate upgrade survivor
