+++ packaging/release_candidate/repository.rb
@@ -79,6 +79,11 @@ module HiveReleaseCandidate
             "packaging/release_candidate/invariant_snapshot.rb",
             "packaging/release_candidate/process_teardown.rb",
             "packaging/release_candidate/upgrade_survivor.rb",
+            "packaging/release_candidate/upgrade_survivor/channel_prefix_oracle.rb",
+            "packaging/release_candidate/upgrade_survivor/reviewed_channel_updater.rb",
+            "packaging/release_candidate/upgrade_survivor/fixed_channel_executor.rb",
+            "packaging/release_candidate/upgrade_survivor/state_snapshotter.rb",
+            "packaging/release_candidate/upgrade_survivor/fixed_phase_executor.rb",
             "packaging/release_candidate/hosted_upgrade_lane.rb",
             "packaging/release_candidate/aggregate.rb",
             "packaging/release_candidate/retry_selection.rb",
